requirements.txt
streamlit
python-dotenv
anthropic
PyYAML
requests
arxiv
openai
Pillow
pylatexenc
PyPDF2
chardet
transformers
torch
pdf2image
pytesseract
PyMuPDF
opencv-python
frontend